quick fix for product editById

+++ src/main/java/com/ashu/boot/services/CategoryServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.ashu.boot.dto.CategoryDTO;
+import com.ashu.boot.exception.CategoryNotFoundException;
 import com.ashu.boot.models.Category;
 import com.ashu.boot.repository.CategoryRepository;
 @Service
@@ -46,28 +47,43 @@ public class CategoryServiceImpl implements CategoryService {
 
 	@Override
 	public CategoryDTO generateOne(Integer id) {
-		Category category= repository.findById(id).get();
-		CategoryDTO categoryDTO = new CategoryDTO();
-		categoryDTO.setId(category.getId());
-		categoryDTO.setName(category.getName());
-		return categoryDTO;
+		Boolean bool =repository.existsById(id);
+		if(bool!=false) {
+			Category category =repository.findById(id).get();
+			CategoryDTO categoryDTO = new CategoryDTO();
+			categoryDTO.setId(category.getId());
+			categoryDTO.setName(category.getName());
+			return categoryDTO;
+		}
+		else {
+			throw new CategoryNotFoundException("CategoryId - "+id+" doesn't exist");
+		}
+		
 	}
 
 	@Override
 	public void edit(int id ,CategoryDTO categoryDTO) {
-		Iterable<Category> categories= repository.findAll();	
-		for(Category category : categories) {
-			if(id ==category.getId()) {
-				category.setName(categoryDTO.getName());
-				repository.save(category);
-			}
+		Boolean bool =repository.existsById(id);
+		if(bool!=false) {
+			Category category =repository.findById(id).get();
+			category.setName(categoryDTO.getName());
+			repository.save(category);	
+		}
+		else {
+			throw new CategoryNotFoundException("CategoryId - "+id+" doesn't exist");
 		}
 	}
 
 	@Override
 	public void del(int id) {
+		Boolean bool =repository.existsById(id);
+		if(bool!=false) {
+			repository.deleteById(id);
+		}
+		else {
+			throw new CategoryNotFoundException("CategoryId - "+id+" doesn't exist");
+		}
 		
-		repository.deleteById(id);
 		
 	}
 }
