+++ src/main/java/com/ashu/boot/services/ProductServiceImpl.java
@@ -7,29 +7,27 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.ashu.boot.dto.ProductDTO;
+import com.ashu.boot.exception.ProductNotFoundException;
 import com.ashu.boot.models.Category;
 import com.ashu.boot.models.Product;
-import com.ashu.boot.repository.CategoryRepository;
 import com.ashu.boot.repository.ProductRepository;
+
 @Service
 public class ProductServiceImpl implements ProductService {
-	
-	
+
 	private ProductRepository repository;
-	private CategoryRepository repositoryCat;
-	
+
 	@Autowired
-	public ProductServiceImpl(ProductRepository repository ,CategoryRepository repositoryCat) {
-	this.repository=repository;	
-	this.repositoryCat=repositoryCat;
+	public ProductServiceImpl(ProductRepository repository) {
+		this.repository = repository;
 	}
-	
+
 	@Override
 	public List<ProductDTO> generate() {
-			 
-		Iterable<Product> products= repository.findAll();
+
+		Iterable<Product> products = repository.findAll();
 		List<ProductDTO> productDTOS = new ArrayList<ProductDTO>();
-		for(Product product : products) {
+		for (Product product : products) {
 			ProductDTO productDTO = new ProductDTO();
 			productDTO.setId(product.getId());
 			productDTO.setName(product.getName());
@@ -37,18 +35,18 @@ public class ProductServiceImpl implements ProductService {
 			productDTO.setPrice(product.getPrice());
 			productDTO.setCategory(product.getCategory());
 			productDTOS.add(productDTO);
-			
+
 		}
-			 return productDTOS;
+		return productDTOS;
 	}
 
 	@Override
-	public Product add(ProductDTO productDTO,int id) {
+	public Product add(ProductDTO productDTO, int id) {
 		Product product = new Product();
 		product.setName(productDTO.getName());
 		product.setQuantity(productDTO.getQuantity());
 		product.setPrice(productDTO.getPrice());
-		productDTO.setCategory(new Category(id,""));
+		productDTO.setCategory(new Category(id, ""));
 		product.setCategory(productDTO.getCategory());
 		repository.save(product);
 		return product;
@@ -56,34 +54,50 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public ProductDTO generateOne(Integer id) {
-		Product product= repository.findById(id).get();
-		ProductDTO productDTO = new ProductDTO();
-		productDTO.setId(product.getId());
-		productDTO.setQuantity(product.getQuantity());
-		productDTO.setName(product.getName());
-		productDTO.setPrice(product.getPrice());
-		productDTO.setCategory(product.getCategory());
-		return productDTO;
+		Boolean bool = repository.existsById(id);
+		if(bool!=false) {
+			Product product = repository.findById(id).get();
+			ProductDTO productDTO = new ProductDTO();
+			productDTO.setId(product.getId());
+			productDTO.setQuantity(product.getQuantity());
+			productDTO.setName(product.getName());
+			productDTO.setPrice(product.getPrice());
+			productDTO.setCategory(product.getCategory());
+			return productDTO;
+			
+		}else {
+			throw new ProductNotFoundException("ProductId - "+id+" doesn't exist" );
+
+		}
+		
 	}
 
 	@Override
-	public void edit(int id ,ProductDTO productDTO,int catId) {
-		Iterable<Product> products=repository.findAll();
-		Category cat=repositoryCat.findById(catId).get();
-		for(Product product : products) {
-			if(id==product.getId()&&cat==product.getCategory()) {
-				product.setName(productDTO.getName());
-				product.setQuantity(productDTO.getQuantity());
-				product.setPrice(productDTO.getPrice());
-				repository.save(product);
-				
-			}
+	public void edit(int Id, ProductDTO productDTO, int categoryId) {
+		Boolean bool = repository.findByIdAndCategoryId(Id, categoryId).isPresent();
+		if (bool != false) {
+			Product product=repository.findByIdAndCategoryId(Id, categoryId).get();
+			product.setName(productDTO.getName());
+			product.setQuantity(productDTO.getQuantity());
+			product.setPrice(productDTO.getPrice());
+			repository.save(product);
 		}
+		else {
+			throw new ProductNotFoundException("ProductId - "+Id+" or Category Id -"+categoryId+ " doesn't exist" );
+		}
+
 	}
 
 	@Override
 	public void del(int id) {
-		repository.deleteById(id);
-		
+		Boolean bool = repository.existsById(id);
+		if (bool != false) {
+			repository.deleteById(id);
+		}
+		else {
+			throw new ProductNotFoundException("ProductId - "+id+"  doesn't exist" );
+		}
+	
+
 	}
 }
